requirements.txt

torch>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
streamlit>=1.27.0
PyPDF2>=3.0.0
python-docx>=0.8.11

numpy >= 1.26
pandas >= 2.2
scikit-learn >= 1.5
transformers >= 4.40


